Updated updateProductt and deleteProductControllers
Updated updateProductt and deleteProductControllers

+++ backend/controllers/product.controller.js
@@ -1,3 +1,4 @@
+import mongoose, { mongo } from 'mongoose';
 import Product  from '../models/product.model.js';
 
 export const getProducts = async(req,res)=>{
@@ -31,6 +32,10 @@ catch(error){
 export const updateProduct =  async (req,res) => {
     const { id } = req.params;
     const product = req.body;
+
+    if(!mongoose.Types.ObjectId.isValid(id)){
+        return res.status(404).json({ success: false, message: "Invalid Product ID"});
+    }
     
     try{
        const updatedProduct =  await Product.findByIdAndUpdate(id,product,{new:true});
@@ -45,12 +50,14 @@ export const updateProduct =  async (req,res) => {
 
 export const deleteProduct = async(req,res) => {
     const { id } = req.params;
-    console.log("id:",id);
+    if(!mongoose.Types.ObjectId.isValid(id)){
+        return res.status(404).json({ success: false, message: "Invalid Product ID"});
+    }
     try{
      await Product.findByIdAndDelete(id);
      res.status(200).json({ success: true, message : "Product Deleted"});
     }catch(error){
         console.log("Error in create Product: ", error.message);
-     res.status(404).json({ success: false, message: "Product not found"});
+     res.status(500).json({ success: false, message: "Server Error"});
     }
 }
